+++ src/dcat-us/_generate-distributions.test.ts
@@ -6,7 +6,7 @@ describe('_generateDistributions', () => {
   const getDownloadLink = (id: string) => `https://my-site.hub.arcgis.com/datasets/${id}`;
 
   it('add default distributions if dataset is a collection', () => {
-    const dataset = {
+    const hubDataset = {
       id: 'foo', // non-layer id
       url: serviceUrl
     };
@@ -27,14 +27,19 @@ describe('_generateDistributions', () => {
     },
   ]
 
-    const distributions = _generateDistributions(dataset, getLandingPage(dataset.id), getDownloadLink(dataset.id));
+    const distributions = _generateDistributions({
+      hubDataset,
+      dcatDataset: {},
+      landingPage: getLandingPage(hubDataset.id), 
+      downloadLink: getDownloadLink(hubDataset.id),
+    });
 
     expect(distributions.length).toBe(2)
     expect(distributions).toEqual(expected);
   })
 
   it('adds default and csv distributions if dataset is a proxied csv', () => {
-    const dataset = {
+    const hubDataset = {
       id: 'foo',
       access: 'public',
       slug: 'nissan::skyline-gtr',
@@ -64,14 +69,19 @@ describe('_generateDistributions', () => {
     }
   ]
 
-    const distributions = _generateDistributions(dataset, getLandingPage(dataset.slug), getDownloadLink(dataset.slug));
+    const distributions = _generateDistributions({
+      hubDataset, 
+      dcatDataset: {},
+      landingPage: getLandingPage(hubDataset.slug), 
+      downloadLink: getDownloadLink(hubDataset.slug)
+    });
 
     expect(distributions.length).toBe(3)
     expect(distributions).toEqual(expected);
   });
 
   it('adds default distributions if dataset is a layer', () => {
-    const dataset = {
+    const hubDataset = {
       id: 'foo_0', // layer id
       url: serviceUrl
     };
@@ -106,14 +116,19 @@ describe('_generateDistributions', () => {
     }
   ]
 
-    const distributions = _generateDistributions(dataset, getLandingPage(dataset.id), getDownloadLink(dataset.id));
+    const distributions = _generateDistributions({
+      hubDataset, 
+      dcatDataset: {},
+      landingPage: getLandingPage(hubDataset.id), 
+      downloadLink: getDownloadLink(hubDataset.id)
+    });
 
     expect(distributions.length).toBe(4)
     expect(distributions).toEqual(expected);
   });
 
   it('adds KML and Shapefile if dataset is a layer AND has geometryType', () => {
-    const dataset = {
+    const hubDataset = {
       id: 'foo_0', // layer id
       url: serviceUrl,
       layer: {
@@ -137,14 +152,19 @@ describe('_generateDistributions', () => {
     }
   ]
 
-    const distributions = _generateDistributions(dataset, getLandingPage(dataset.id), getDownloadLink(dataset.id));
+    const distributions = _generateDistributions({
+      hubDataset, 
+      dcatDataset: {},
+      landingPage: getLandingPage(hubDataset.id), 
+      downloadLink: getDownloadLink(hubDataset.id)
+    });
 
     expect(distributions.length).toBe(6)
     expect(distributions.slice(4)).toEqual(expected);
   });
 
   it('adds WFS and WMS if supported', () => {
-    const dataset = {
+    const hubDataset = {
       id: 'foo', // non-layer id
       url: serviceUrl
     };
@@ -165,21 +185,36 @@ describe('_generateDistributions', () => {
       accessURL: 'https://servicesqa.arcgis.com/Xj56SBi2udA78cC9/arcgis/services/Tahoe_Things/FeatureServer/WFSServer?request=GetCapabilities&service=WFS'
     };
 
-    const justWFS = _generateDistributions({ ...dataset, supportedExtensions: 'WFSServer' }, getLandingPage(dataset.id), getDownloadLink(dataset.id));
+    const justWFS = _generateDistributions({
+      hubDataset: { ...hubDataset, supportedExtensions: 'WFSServer' }, 
+      dcatDataset: {},
+      landingPage: getLandingPage(hubDataset.id), 
+      downloadLink: getDownloadLink(hubDataset.id)
+    });
     expect(justWFS.length).toBe(3);
     expect(justWFS.pop()).toEqual(expectedWFSDistribution);
 
-    const justWMS = _generateDistributions({ ...dataset, supportedExtensions: 'WMSServer' }, getLandingPage(dataset.id), getDownloadLink(dataset.id));
+    const justWMS = _generateDistributions({
+      hubDataset: { ...hubDataset, supportedExtensions: 'WMSServer' }, 
+      dcatDataset: {},
+      landingPage: getLandingPage(hubDataset.id), 
+      downloadLink: getDownloadLink(hubDataset.id)
+    });
     expect(justWMS.length).toBe(3);
     expect(justWMS.pop()).toEqual(expectedWMSDistribution);
 
-    const allOGCServices = _generateDistributions({ ...dataset, supportedExtensions: 'WMSServer,WFSServer' }, getLandingPage(dataset.id), getDownloadLink(dataset.id));
+    const allOGCServices = _generateDistributions({
+      hubDataset: { ...hubDataset, supportedExtensions: 'WMSServer,WFSServer' }, 
+      dcatDataset: {},
+      landingPage: getLandingPage(hubDataset.id), 
+      downloadLink: getDownloadLink(hubDataset.id)
+    });
     expect(allOGCServices.length).toBe(4);
     expect(allOGCServices.slice(2)).toEqual([expectedWFSDistribution, expectedWMSDistribution]);
   });
 
-  it('adds custom distributions from metadata', () => {
-    const dataset = {
+  it('adds distributions from metadata', () => {
+    const hubDataset = {
       id: 'foo', // non-layer id
       url: serviceUrl,
       metadata: {
@@ -206,7 +241,12 @@ describe('_generateDistributions', () => {
       }
     };
 
-    const distributions = _generateDistributions(dataset, getLandingPage(dataset.id), getDownloadLink(dataset.id));
+    const distributions = _generateDistributions({
+      hubDataset, 
+      dcatDataset: {},
+      landingPage: getLandingPage(hubDataset.id), 
+      downloadLink: getDownloadLink(hubDataset.id)
+    });
 
     expect(distributions.length).toBe(4);
     expect(distributions.slice(2)).toEqual([
@@ -226,4 +266,82 @@ describe('_generateDistributions', () => {
       }
     ])
   });
+
+  it('does not add custom distributions if interpolated value is an empty array', () => {
+    const hubDataset = {
+      id: 'foo', // non-layer id
+      url: serviceUrl
+    };
+
+    const dcatDataset = {
+      distribution: []
+    }
+
+    const expected =  [
+    {
+      '@type': 'dcat:Distribution',
+      title: 'ArcGIS Hub Dataset',
+      format: 'Web Page',
+      mediaType: 'text/html',
+      accessURL: 'https://my-site.hub.arcgis.com/maps/foo'
+    },
+    {
+      '@type': 'dcat:Distribution',
+      title: 'ArcGIS GeoService',
+      format: 'ArcGIS GeoServices REST API',
+      mediaType: 'application/json',
+      accessURL: 'https://servicesqa.arcgis.com/Xj56SBi2udA78cC9/arcgis/rest/services/Tahoe_Things/FeatureServer/0'
+    },
+  ]
+
+    const distributions = _generateDistributions({
+      hubDataset,
+      dcatDataset,
+      landingPage: getLandingPage(hubDataset.id), 
+      downloadLink: getDownloadLink(hubDataset.id),
+    });
+
+    expect(distributions.length).toBe(2)
+    expect(distributions).toEqual(expected);
+  });
+
+  it('does not add custom distributions if interpolated value is an object', () => {
+    const hubDataset = {
+      id: 'foo', // non-layer id
+      url: serviceUrl
+    };
+
+    const dcatDataset = {
+      distribution: {
+        customKey: 'customValue'
+      }
+    }
+
+    const expected =  [
+    {
+      '@type': 'dcat:Distribution',
+      title: 'ArcGIS Hub Dataset',
+      format: 'Web Page',
+      mediaType: 'text/html',
+      accessURL: 'https://my-site.hub.arcgis.com/maps/foo'
+    },
+    {
+      '@type': 'dcat:Distribution',
+      title: 'ArcGIS GeoService',
+      format: 'ArcGIS GeoServices REST API',
+      mediaType: 'application/json',
+      accessURL: 'https://servicesqa.arcgis.com/Xj56SBi2udA78cC9/arcgis/rest/services/Tahoe_Things/FeatureServer/0'
+    },
+  ]
+
+    const distributions = _generateDistributions({
+      hubDataset,
+      dcatDataset,
+      landingPage: getLandingPage(hubDataset.id), 
+      downloadLink: getDownloadLink(hubDataset.id),
+    });
+
+    expect(distributions.length).toBe(2)
+    expect(distributions).toEqual(expected);
+  })
 });