+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # CHANGELOG.md
 
+## Unreleased
+Added
+- If the `distribution` property of a site's dcat config is an array, those custom distributions will now be prepended to the distributions list [#23](https://github.com/koopjs/koop-output-dcat-us-11/pull/23)
+
 ## 1.8.2
 Fixed
 - Fixed access urls for downloadable distributions [#22](https://github.com/koopjs/koop-output-dcat-us-11/pull/22)